Move login activities to a login package
