+++ .idea/misc.xml
@@ -23,6 +23,7 @@
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.259375" />
         <entry key="app/src/main/res/layout/activity_main_screen.xml" value="0.259375" />
         <entry key="app/src/main/res/layout/activity_register_email.xml" value="0.35625" />
+        <entry key="app/src/main/res/layout/activity_verify_email.xml" value="0.259375" />
         <entry key="app/src/main/res/layout/calendar_row.xml" value="0.25317028985507245" />
         <entry key="app/src/main/res/layout/calendar_upcoming_row.xml" value="0.20260416666666667" />
         <entry key="app/src/main/res/layout/calendar_weekly_row.xml" value="0.1890625" />
@@ -30,6 +31,9 @@
         <entry key="app/src/main/res/layout/fragment_first.xml" value="0.2210144927536232" />
         <entry key="app/src/main/res/layout/fragment_second.xml" value="0.2210144927536232" />
         <entry key="app/src/main/res/layout/house_row.xml" value="0.22" />
+        <entry key="app/src/main/res/layout/subtask_row.xml" value="0.25" />
+        <entry key="app/src/main/res/layout/task_editor.xml" value="0.22" />
+        <entry key="app/src/main/res/layout/task_row.xml" value="0.12183383991894631" />
       </map>
     </option>
   </component>
