+++ app/build.gradle
@@ -48,7 +48,7 @@ dependencies {
     implementation 'com.firebaseui:firebase-ui-firestore:8.0.0'
 
     // Import the BoM for the Firebase platform
-    implementation platform('com.google.firebase:firebase-bom:29.2.0')
+    implementation platform('com.google.firebase:firebase-bom:29.2.1')
 
     implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation 'com.google.android.material:material:1.5.0'
