+++ app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.HouseOrganizer">
         <activity
-            android:name=".VerifyEmail"
+            android:name=".login.VerifyEmail"
             android:exported="false" />
         <activity
-            android:name=".RegisterEmail"
+            android:name=".login.RegisterEmail"
             android:exported="false" />
         <activity
             android:name=".CreateHouseholdActivity"
@@ -24,7 +24,7 @@
             android:label="@string/title_activity_settings"
             android:parentActivityName=".MainScreenActivity" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".login.LoginActivity"
             android:exported="false" />
         <activity
             android:name=".HouseSelectionActivity"
@@ -34,7 +34,7 @@
             android:name=".MainScreenActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginEmail"
+            android:name=".login.LoginEmail"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
