+++ app/src/main/res/layout/activity_login_email.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginEmail">
+    tools:context=".login.LoginEmail">
 
     <Button
         android:id="@+id/log_email_signin_button"
