+++ app/src/main/res/layout/activity_register_email.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegisterEmail">
+    tools:context=".login.RegisterEmail">
 
     <EditText
         android:id="@+id/reg_confirm_password"
