+++ app/src/main/res/layout/activity_verify_email.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".VerifyEmail">
+    tools:context=".login.VerifyEmail">
 
     <TextView
         android:id="@+id/verif_email_text"
